add clean examples and fix react warning
